factor: Move domain folder  to internal
