+++ go/services/users/internal/domain/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 import (
-	valueobject "GROWTHSPHERE/services/users/domain/valueObject"
+	valueobject "GROWTHSPHERE/services/users/internal/domain/valueObject"
 	"time"
 
 	"github.com/google/uuid"
